LayoutTests/fast/events/touch/touch-event-constructor-expected.txt

Test the TouchEvent constructor

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ev = new TouchEvent('touchstart') did not throw exception.
PASS ev.__proto__ is TouchEvent.prototype
PASS ev.type is "touchstart"
PASS ev.touches.length is 0
PASS ev.targetTouches.length is 0
PASS ev.changedTouches.length is 0
PASS ev.bubbles is false
PASS ev = new TouchEvent('touchmove', { touches: listA, bubbles: true }) did not throw exception.
PASS ev.__proto__ is TouchEvent.prototype
PASS ev.type is "touchmove"
PASS ev.touches.length is 1
PASS ev.targetTouches.length is 0
PASS ev.changedTouches.length is 0
PASS ev.bubbles is true
PASS ev = new TouchEvent('touchmove', { targetTouches: listB, bubbles: true }) did not throw exception.
PASS ev.__proto__ is TouchEvent.prototype
PASS ev.type is "touchmove"
PASS ev.touches.length is 0
PASS ev.targetTouches.length is 2
PASS ev.changedTouches.length is 0
PASS ev.bubbles is true
PASS ev = new TouchEvent('touchmove', { changedTouches: listC, bubbles: true }) did not throw exception.
PASS ev.__proto__ is TouchEvent.prototype
PASS ev.type is "touchmove"
PASS ev.touches.length is 0
PASS ev.targetTouches.length is 0
PASS ev.changedTouches.length is 3
PASS ev.bubbles is true
PASS ev = new TouchEvent('touchmove', { touches: listA, targetTouches: listB, changedTouches: listC, bubbles: true }) did not throw exception.
PASS ev.__proto__ is TouchEvent.prototype
PASS ev.type is "touchmove"
PASS ev.touches.length is 1
PASS ev.targetTouches.length is 2
PASS ev.changedTouches.length is 3
PASS ev.bubbles is true
PASS successfullyParsed is true

TEST COMPLETE
